requirements.txt
pyTelegramBotAPI==4.25.0
Flask==3.0.3
python-dotenv==1.0.1
